chore: Migrating interface declaration to global declaration file and fix linting
