+++ package.json
@@ -39,6 +39,7 @@
         "commitlint": "9.1.2",
         "eslint": "7.8.1",
         "eslint-config-react-app": "5.2.1",
+        "eslint-config-recommended": "4.1.0",
         "eslint-plugin-flowtype": "5.2.0",
         "eslint-plugin-import": "2.22.0",
         "eslint-plugin-jest": "23.20.0",
