--- packages/server/src/interfaces/dataStoredInToken.d.ts
@@ -1,5 +1,0 @@
-interface DataStoredInToken {
-  _id: string;
-}
-
-export default DataStoredInToken;
