--- packages/server/src/interfaces/requestWithUser.interface.d.ts
@@ -1,8 +1,0 @@
-import { Request } from 'express';
-import User from 'users/user.interface';
-
-interface RequestWithUser extends Request {
-  user: User;
-}
-
-export default RequestWithUser;
