--- packages/server/src/interfaces/tokenData.interface.d.ts
@@ -1,6 +1,0 @@
-interface TokenData {
-  token: string;
-  expiresIn: number;
-}
-
-export default TokenData;
