--- packages/server/src/interfaces/userCreationData.d.ts
@@ -1,6 +1,0 @@
-type UserCreationData = {
-    cookie: string;
-    user: User;
-}
-
-export default UserCreationData;
