+++ packages/server/src/middleware/auth.middleware.ts
@@ -2,16 +2,14 @@ import { NextFunction, Response } from 'express';
 import jwt from 'jsonwebtoken';
 import AuthenticationTokenMissingException from '../exceptions/AuthenticationTokenMissingException';
 import WrongAuthenticationTokenException from '../exceptions/WrongAuthenticationTokenException';
-import DataStoredInToken from '../interfaces/dataStoredInToken';
-import RequestWithUser from '../interfaces/requestWithUser.interface';
 import userModel from '../user/user.model';
 
-async function authMiddleware(request: RequestWithUser, response: Response, next: NextFunction): Promise<void> {
+async function authMiddleware(request: Moai.RequestWithUser, response: Response, next: NextFunction): Promise<void> {
     const cookies = request.cookies;
     if (cookies && cookies.Authorization) {
         const secret = process.env.JWT_SECRET;
         try {
-            const verificationResponse = jwt.verify(cookies.Authorization, secret) as DataStoredInToken;
+            const verificationResponse = jwt.verify(cookies.Authorization, secret) as Moai.DataStoredInToken;
             const id = verificationResponse._id;
             const user = await userModel.findById(id);
             if (user) {
